+++ go/CHANGELOG.md
@@ -10,6 +10,7 @@
 - Preserve external log message order (PR: keybase/client#1641)
 - SaltPack: implement sign/verify commands (PR: keybase/client#1635)
 - SaltPack: implement sign/verify package (PRs: keybase/client#1596, keybase/client#1612, keybase/client#1614)
+- SaltPack: implement the sender secretbox (PR: keybase/client#1645)
 - Fix merkle tree path mismatch bug (PR: keybase/client#1621)
 - encoding: Speed up B62 decoder (PRs: keybase/client#1644, keybase/client#1640)